refactor: theme folder 구조 변경, types를 palette 파일로 병합(#79)
