+++ packages/react/components/Loader/Loaders/Bars/Bars.styles.ts
@@ -1,7 +1,7 @@
 import { css, keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
 import { palette } from '../../../../theme/palette';
-import { NormalColorType } from '../../../../theme/types';
+import { NormalColorType } from '../../../../theme';
 
 const BarKeyframes = keyframes`
   0% {
