+++ packages/react/components/Loader/Loaders/Spinner/Spinner.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { NormalColorType } from '../../../../theme/types';
+import { NormalColorType } from '../../../../theme';
 import * as S from './Spinner.styles';
 
 interface SpinnerProps {
