+++ packages/react/index.ts
@@ -1,5 +1,4 @@
 export * from './components';
 export * from './layouts';
 export * from './hooks';
-
-export { WapUIProvider } from './theme/theme-provider';
+export * from './theme';
